Add files via upload
Java application providing annotated visualizations of MSMS spectra stored in Peptide Depot. Can be run as an applet to provide an interactive view of a single spectra OR as a GUI-less application to generate annotated PDFs of all the MSMS spectra from an LC-MS run.

To run as an applet, the code is compiled using the main class in vmsdb, and vmsdb_pdf is NOT used.
To run as a GUI-less PDF maker, the code is compiled using the main class in vmsdb_pdf, and vmsdb is NOT used.